web/requirements.txt
Django==4.2.7
gunicorn==20.1.0
django-prometheus==2.2.0
psycopg2-binary==2.9.7
python-decouple==3.6
black==23.7.0
flake8==6.1.0
Pillow==11.0.0
cryptography==44.0.0